+++ ext/posix/sys/msg.c
@@ -33,6 +33,122 @@
 #include <sys/msg.h>
 #include <sys/types.h>
 
+
+/***
+Message queue record.
+@table PosixMsqid
+@int msg_qnum number of messages on the queue
+@int msg_qbytes number of bytes allowed on the queue
+@int msg_lspid process id of last msgsnd
+@int msg_lrpid process id of last msgrcv
+@int msg_stime time of last msgsnd
+@int msg_rtime time of last msgrcv
+@int msg_ctime time of last change
+*/
+static int
+pushmsqid(lua_State *L, struct msqid_ds *msqid)
+{
+	if (!msqid)
+		return lua_pushnil(L), 1;
+
+	lua_createtable(L, 0, 8);
+	setintegerfield(msqid, msg_qnum);
+	setintegerfield(msqid, msg_qbytes);
+	setintegerfield(msqid, msg_lspid);
+	setintegerfield(msqid, msg_lrpid);
+	setintegerfield(msqid, msg_stime);
+	setintegerfield(msqid, msg_rtime);
+	setintegerfield(msqid, msg_ctime);
+
+	lua_createtable(L, 0, 5);
+	pushintegerfield("uid", msqid->msg_perm.uid);
+	pushintegerfield("gid", msqid->msg_perm.gid);
+	pushintegerfield("cuid", msqid->msg_perm.cuid);
+	pushintegerfield("cgid", msqid->msg_perm.cgid);
+	pushintegerfield("mode", msqid->msg_perm.mode);
+
+	lua_setfield(L, -2, "msg_perm");
+	settypemetatable("PosixMsqid");
+	return 1;
+}
+
+
+static const char *Smsqid_fields[] = { "msg_qbytes", "msg_perm" };
+
+
+static const char *Sipcperm_fields[] = { "uid", "gid", "mode" };
+
+
+static void
+tomsqid(lua_State *L, int index, struct msqid_ds *msqid)
+{
+	int subindex;
+	luaL_checktype(L, index, LUA_TTABLE);
+
+	/* Copy fields to msqid struct */
+	msqid->msg_qbytes = checkintfield(L, index, "msg_qbytes");
+
+	checkfieldtype(L, index, "msg_perm", LUA_TTABLE, "table");
+	subindex = lua_gettop(L);
+        msqid->msg_perm.uid  = checkintfield(L, subindex, "uid");
+	msqid->msg_perm.gid  = checkintfield(L, subindex, "gid");
+	msqid->msg_perm.mode = checkintfield(L, subindex, "mode");
+
+        checkfieldnames(L, index, Smsqid_fields);
+	checkfieldnames(L, subindex, Sipcperm_fields);
+}
+
+
+
+/***
+@function msgctl
+@int id message queue identifier returned by @{msgget}
+@int cmd one of `IPC_STAT`, `IPC_SET` or `IPC_RMID`
+@PosixMsqid[opt=nil] values to set with `IPC_SET`
+@treturn[1] PosixMsqid table for *id*, with `IPC_STAT`, if successful
+@treturn[2] non-nil, with `IPC_SET` or `IPC_RMID`, if successful
+@treturn[3] nil otherwise
+@treturn[3] string error message
+@treturn[3] int errnum
+@see msgctl(2)
+@usage
+  local sysvmsg = require 'posix.sys.msg'
+  local msq = sysvmsg.msgget(sysvmsg.IPC_PRIVATE)
+  local msqid, errmsg = sysvmsg.msgctl(msq, sysvmsg.IPC_STAT)
+  assert(msqid, errmsg)
+  assert(sysvmsg.msgctl(msq, sysvmsg.IPC_RMID))
+*/
+static int
+Pmsgctl(lua_State *L)
+{
+	int id = checkint(L, 1);
+	int cmd = checkint(L, 2);
+	struct msqid_ds msqid;
+
+	switch (cmd)
+	{
+		case IPC_RMID:
+			checknargs(L, 2);
+			return pushresult(L, msgctl(id, cmd, NULL), "msgctl");
+
+		case IPC_SET:
+			checknargs(L, 3);
+			tomsqid(L, 3, &msqid);
+			return pushresult(L, msgctl(id, cmd, &msqid), "msgctl");
+
+		case IPC_STAT:
+			checknargs(L, 2);
+			if (msgctl(id, cmd, &msqid) < 0)
+				return pusherror(L, "msgctl");
+			return pushmsqid(L, &msqid);
+
+		default:
+			checknargs(L, 3);
+			return pusherror(L, "unsupported cmd value");
+	}
+}
+
+
 /***
 Get a message queue identifier
 @function msgget
@@ -156,6 +272,7 @@ Pmsgrcv(lua_State *L)
 static const luaL_Reg posix_sys_msg_fns[] =
 {
 #if HAVE_SYSV_MESSAGING
+	LPOSIX_FUNC( Pmsgctl		),
 	LPOSIX_FUNC( Pmsgget		),
 	LPOSIX_FUNC( Pmsgsnd		),
 	LPOSIX_FUNC( Pmsgrcv		),
@@ -173,6 +290,9 @@ Constants.
 Message constants.
 Any constants not available in the underlying system will be `nil` valued.
 @table posix.sys.msg
+@int IPC_STAT return a Msqid table from msgctl
+@int IPC_SET set the Msqid fields from msgctl
+@int IPC_RMID remove a message queue with msgctl
 @int IPC_CREAT create entry if key does not exist
 @int IPC_EXCL fail if key exists
 @int IPC_PRIVATE private key
@@ -197,15 +317,18 @@ luaopen_posix_sys_msg(lua_State *L)
 
 #if HAVE_SYSV_MESSAGING
 	LPOSIX_CONST( IPC_CREAT		);
-	LPOSIX_CONST( IPC_EXCL		);
-	LPOSIX_CONST( IPC_PRIVATE	);
-	LPOSIX_CONST( IPC_NOWAIT	);
 #  ifdef MSG_EXCEPT
 	LPOSIX_CONST( MSG_EXCEPT	);
 #  endif
+	LPOSIX_CONST( IPC_EXCL		);
 #  ifdef MSG_NOERROR
 	LPOSIX_CONST( MSG_NOERROR	);
 #  endif
+	LPOSIX_CONST( IPC_NOWAIT	);
+	LPOSIX_CONST( IPC_PRIVATE	);
+	LPOSIX_CONST( IPC_RMID		);
+	LPOSIX_CONST( IPC_SET		);
+	LPOSIX_CONST( IPC_STAT		);
 #endif
 
 	return 1;